Fix for RESTX-42: No more hard-wired component names in a Python source file. Instead, we now have small 'manifest' files (not in XML) for each component. The server reads those files at start and dynamically imports the necessary component classes.
The restxctl script is updated to now just create a manifest file.

Also, it's possible to now 'create' a new component for which we have the
source file already. In that case, only the manifest is created, the source
template file is NOT put in place of the old component code.