Added a self-driven demo.
Lots of static HTML files, CSS and JS for the frame-based demo.
Various modifications to the rendering of component and resource
data (adding 'id' attributes, for example), so that the JS of the
demo can work well with it.
When no input types are wanted then I changed to 'None' (displaying
'---' next to input_types).